Fixes Ticket #168 - Stops rat crashing when two (or more) instances run in encrypted mode using different keys on the same address with promiscuous mode switched on.
Promiscuous mode (where RAT doesn't wait for an RTCP packet before allowing the source - ie it only requires a RTP packet and omits the 'probation' check for in sequence RTP packets) has been on as default for a while as it useful for quickly connecting to devices that don't generate RTCP. However promiscuous mode is too relaxed about packet admission. Despite passing the rtp_validate() function some packets - especially those from a 'wrongly' (the key isn't correct) decrypted session - manage to pass that check but are actually garbage. The additional 'probation' checks in the promiscuous code (which checked for more than one packet in the sequence) are switched off. 

Solution:
- The wait_for_rtcp rtp option has had it's token renamed from RTP_OPT_WEAK_VALIDATION to RTP_OPT_WAIT_FOR_RTCP - It's value has been switched for previous usage as it makes more sense with the new naming. Code has been cleaned up. 
- Basically the fixes lies in the changed default config: RAT now enabled wait_for_rtcp and disables promiscuous mode.