Bumped version number to make it easier to enforce RPM and .deb dependecies for packages based on ANU IF's VPMedia which originally required ANU IF common.
Andrew Ford noticed there was a bit of a performance regression with HD video after he switched VPMedia to link to the UCL version of the common library - turned out it was setting the buffer size too.

Applied patch from Andrew which increased UDP buffer size to to match what the if-media common had. So far no latenency issues have been found, but need to double check RAT.